+++ app.config.ts
@@ -54,22 +54,31 @@ const config: ExpoConfig = {
   extra: {
     authLoginPage: process.env.AUTH_LOGIN_PAGE,
     eas: {
-      projectId: 'e6c8b8ff-a5de-4c93-896a-21e3962ff7a5',
+      projectId: 'eb428ba3-923b-4b72-bdcd-7ab92d0e0321',
     },
   },
   icon: './assets/icon.png',
   ios: {
     buildNumber: getBuildNumber(version),
-    bundleIdentifier: 'blue10.InvoiceManagement',
+    bundleIdentifier: 'com.blue10.Scanner',
+    config: {
+      usesNonExemptEncryption: false,
+    },
+    infoPlist: {
+      CFBundleLocalizations: ['en', 'nl'],
+    },
     supportsTablet: true,
   },
   name: 'Blue10',
   orientation: 'portrait',
-  owner: 'wecreatesolutions',
+  owner: 'blue10',
   plugins: [
     'sentry-expo',
+    ['react-native-document-scanner-plugin', {
+      cameraPermission: 'To scan documents, camera access is required.',
+    }],
   ],
-  slug: 'blue10',
+  slug: 'blue10-app',
   splash: {
     backgroundColor: '#333333',
     image: './assets/splash.png',