+++ src/i18n/types.ts
@@ -2,6 +2,7 @@ export type TranslationKeysType = {
   general: {
     error: string;
     button_ok: string;
+    unknown: string;
   };
   login: {
     welcome_title: string;
@@ -117,5 +118,35 @@ export type TranslationKeysType = {
     title: string;
     description: string;
     button_retry: string;
+  };
+  scan: {
+    company_title: string;
+    document_type_packing_slip: string;
+    document_type_purchase_invoice: string;
+    document_type_sales_invoice: string;
+    document_type_title: string;
+    document_type_undefined: string;
+    menu_cancel: string;
+    menu_dashboard: string;
+    menu_select_company: string;
+    menu_select_document_type: string;
+    menu_title: string;
+    preview_title: string;
+    preview_upload_failed_title: string;
+    preview_upload_success_title: string;
+    screen_title: string;
+    upload_busy_toast: string;
+    upload_finalize_warning: string;
+    upload_generate_document_error: string;
+    upload_start_session_error: string;
+    upload_success: string;
+    upload_table_company: string;
+    upload_table_customer: string;
+    upload_table_document_type: string;
+    upload_table_images: string;
+    upload_table_user: string;
+    upload_title: string;
+    upload_upload_error: string;
+    upload_user_aborted: string;
   };
 };