+++ src/screens/DashboardScreen.tsx
@@ -1,9 +1,10 @@
 import { NativeStackScreenProps } from '@react-navigation/native-stack';
-import React from 'react';
+import React, { useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { StyleSheet, View } from 'react-native';
 
 // import SvgSwitchIcon from '../../assets/icons/dashboard-switch-icon.svg';
+import SvgCameraShape from '../../assets/icons/camerashape.svg';
 import SvgGearShape from '../../assets/icons/gearshape.svg';
 import { DashboardItem } from '../components/DashboardItem/DashboardItem';
 import LoaderWrapper from '../components/LoaderWrapper/LoaderWrapper';
@@ -13,6 +14,7 @@ import { useGetApprovedInvoiceCount } from '../hooks/queries/useGetApproveInvoic
 import { useGetCurrentCustomer } from '../hooks/queries/useGetCurrentCustomer';
 import { useGetCurrentUser } from '../hooks/queries/useGetCurrentUser';
 import { RootStackParamList } from '../navigation/types';
+import { useImageStore } from '../store/ImageStore';
 import { colors, dimensions, text } from '../theme/';
 
 type Props = NativeStackScreenProps<RootStackParamList, 'Dashboard'>;
@@ -21,6 +23,7 @@ export const DashboardScreen: React.FC<Props> = ({ navigation }) => {
   const { currentUser, query: { isLoading } } = useGetCurrentUser();
   const currentCustomer = useGetCurrentCustomer();
   const { count: invoices, isLoading: isCountLoading } = useGetApprovedInvoiceCount();
+  const { reset: resetScannedImages } = useImageStore();
   const { t } = useTranslation();
 
   // const onSwitchEnv = () => {
@@ -35,6 +38,19 @@ export const DashboardScreen: React.FC<Props> = ({ navigation }) => {
       },
     );
   };
+
+  const canUserUploadOrApprove = useMemo(
+    () => currentUser?.MayHandleAllCompanies === true ||
+      currentUser?.MayApproveAllCompanies === true ||
+      (currentUser?.HandleCompanies.length || 0) > 0 ||
+      (currentUser?.ApproveCompanies.length || 0) > 0,
+    [currentUser],
+  );
+
+  const startScannerFlow = useCallback(() => {
+    resetScannedImages();
+    navigation.navigate('ScanSelectCompanyScreen');
+  }, [navigation, resetScannedImages]);
 
   return (
     <ScreenWithStatusBarAndHeader>
@@ -59,14 +75,27 @@ export const DashboardScreen: React.FC<Props> = ({ navigation }) => {
           <SvgSwitchIcon style={{ alignSelf: 'center' }} />
         </DashboardItem>
         */}
-        <DashboardItem
-          isLoading={isCountLoading}
-          title={t('to_approved_invoices.screen_title')}
-          color={colors.dashboard.approval.background}
-          textColor={colors.dashboard.approval.text}
-          contentTitle={invoices.toString()}
-          onPress={onApproveInvoices}
-        />
+        {canUserUploadOrApprove && (
+          <DashboardItem
+            isLoading={false}
+            title={t('scan.screen_title')}
+            color={colors.dashboard.scan.background}
+            textColor={colors.dashboard.scan.text}
+            onPress={startScannerFlow}
+          >
+            <SvgCameraShape color={colors.white} style={{ alignSelf: 'center' }} width={75} height={75} />
+          </DashboardItem>
+        )}
+        {canUserUploadOrApprove && (
+          <DashboardItem
+            isLoading={isCountLoading}
+            title={t('to_approved_invoices.screen_title')}
+            color={colors.dashboard.approval.background}
+            textColor={colors.dashboard.approval.text}
+            contentTitle={invoices.toString()}
+            onPress={onApproveInvoices}
+          />
+        )}
         <DashboardItem
           isLoading={false}
           title={t('settings.screen_title')}