+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v0.4.0
+* \#42 Dashboard Page knop toevoegen scannen
+* \#43 Scannen van een document
+* \#44 Uploaden van een scan
+
 # v0.3.0
 * \#54 Inloggen via LoginSite
 * \#56 Opmerkingen boven de actieknoppen plaatsen op een factuur.